concurrent heterogeneous queue (WIP)

+++ density_tests/functionality_cases/concurrent_producer_consumer_test.h
@@ -23,19 +23,19 @@ namespace density_tests
         ConcProdConsTest & operator = (const ConcProdConsTest&) = delete;
 
         /* If i_consumer_count and are both zero, the test runs on a single thread */
-        ConcProdConsTest(size_t i_consumer_count, size_t i_producer_count, size_t i_cell_count)
+        ConcProdConsTest(int64_t i_consumer_count, int64_t i_producer_count, int64_t i_cell_count)
             : m_produced(0), m_consumed(0), m_cell_count(i_cell_count)
         {
             m_cells = std::make_unique<sync::atomic<uint8_t>[]>(i_cell_count);
-            for (size_t i = 0; i < i_cell_count; i++)
+            for (int64_t i = 0; i < i_cell_count; i++)
                 m_cells[i].store(0);
 
-            for (size_t i = 0; i < i_consumer_count; i++)
+            for (int64_t i = 0; i < i_consumer_count; i++)
             {
                 m_threads.emplace_back([=] { consumer_procedure(i_cell_count); });
             }
 
-            for (size_t i = 0; i < i_producer_count; i++)
+            for (int64_t i = 0; i < i_producer_count; i++)
             {
                 m_threads.emplace_back([=] { producer_procedure(i, i_cell_count, i_producer_count); });
             }
@@ -62,7 +62,7 @@ namespace density_tests
                 thread.join();
             }
 
-            for (size_t i = 0; i < m_cell_count; i++)
+            for (int64_t i = 0; i < m_cell_count; i++)
             {
                 auto val = m_cells[i].load();
                 TESTITY_ASSERT(val == 2);
@@ -71,29 +71,29 @@ namespace density_tests
 
     private:
 
-        void producer_procedure(size_t i_from, size_t i_to, size_t i_step)
+        void producer_procedure(int64_t i_from, int64_t i_to, int64_t i_step)
         {
-            for (size_t i = i_from; i < i_to; i += i_step)
+            for (int64_t i = i_from; i < i_to; i += i_step)
             {
                 auto prev_val = m_cells[i].fetch_add(1);
                 TESTITY_ASSERT(prev_val == 0);
-                m_container.push((uint64_t)i);
+                m_container.push((int64_t)i);
                 ++m_produced;
             }
         }
 
-        void consumer_procedure(size_t i_cell_count)
+        void consumer_procedure(int64_t i_cell_count)
         {
-            auto const int_type = CONTAINER::runtime_type::template make<uint64_t>();
+            auto const int_type = CONTAINER::runtime_type::template make<int64_t>();
 
             while (m_consumed.load() < i_cell_count)
             {
                 bool res = m_container.try_consume([this, &int_type](const typename CONTAINER::runtime_type & i_complete_type, void * i_element) {
                     TESTITY_ASSERT(i_complete_type == int_type);
 
-                    auto value = *static_cast<uint64_t*>(i_element);
+                    auto value = *static_cast<int64_t*>(i_element);
 
-                    *static_cast<uint64_t*>(i_element) = ~*static_cast<uint64_t*>(i_element);
+                    *static_cast<int64_t*>(i_element) = ~*static_cast<int64_t*>(i_element);
 
                     auto prev_val = m_cells[value].fetch_add(1);
                     TESTITY_ASSERT(prev_val == 1);
@@ -110,9 +110,9 @@ namespace density_tests
 
     private:
         std::vector<std::thread> m_threads;
-        sync::atomic<uint64_t> m_produced, m_consumed;
+        sync::atomic<int64_t> m_produced, m_consumed;
         CONTAINER m_container;
         std::unique_ptr<sync::atomic<uint8_t>[]> m_cells;
-        const size_t m_cell_count;
+        const int64_t m_cell_count;
     };
 }
